requirements.txt
flask
gunicorn
pyyaml
pillow
watchdog
Flask-PyMongo
